Rename ResponsesController to FormsController
